fix: move sendable to module
