+++ crates/provider/src/provider/mod.rs
@@ -1,5 +1,8 @@
 mod r#trait;
-pub use r#trait::{FilterPollerBuilder, Provider, SendableTx};
+pub use r#trait::{FilterPollerBuilder, Provider};
 
 mod root;
 pub use root::RootProvider;
+
+mod sendable;
+pub use sendable::SendableTx;
